Revert "Upgrading of test facilities (#29)"
This reverts commit fa76198522ee16df45d3d390041b852c6987c6c4.
